refactor: [route] 프로젝트 구조 refactoring
